--- browser_use/requirements.txt
@@ -1,17 +1,0 @@
-langchain-core~=0.3.49
-langchain-google-genai~=2.1.2
-langchain-openai~=0.3.11
-pydantic~=2.10.6
-httpx~=0.28.1
-boto3~=1.38.27
-botocore~=1.38.27
-langchain-aws~=0.2.19
-protobuf~=6.31.1
-beautifulsoup4~=4.13.4
-pyperclip~=1.9.0
-openai~=1.82.1
-uuid7~=0.1.0
-click~=8.2.1
-psutil~=7.0.0
-camoufox~=0.4.11
-posthog~=3.25.0
