+++ requirements.txt
@@ -8,10 +8,14 @@ langchain-core~=0.3.49
 langchain-google-genai~=2.1.2
 langchain-openai~=0.3.11
 pydantic~=2.10.6
+playwright~=1.52.0
 httpx~=0.28.1
+rich~=14.0.0
 boto3~=1.38.27
 botocore~=1.38.27
 langchain-aws~=0.2.19
+patchright~=1.52.4
+aiohttp~=3.12.6
 protobuf~=6.31.1
 beautifulsoup4~=4.13.4
 pyperclip~=1.9.0
@@ -20,9 +24,5 @@ uuid7~=0.1.0
 click~=8.2.1
 psutil~=7.0.0
 camoufox~=0.4.11
-posthog~=3.25.0
-playwright-recaptcha~=0.5.0
 capsolver~=1.0.0
-pydub~=0.25.1
-SpeechRecognition~=3.10.0
-tenacity~=8.2.0
+posthog~=3.25.0
