+++ tests/undetected_test.py
@@ -3,7 +3,7 @@ import sys
 
 from dotenv import load_dotenv
 load_dotenv()
-from browser_use import StealthAgent, utils
+from browser_use import StealthAgent, stealth_utils
 from langchain_openai import ChatOpenAI
 
 async def main():
@@ -11,7 +11,7 @@ async def main():
         task="Find the price difference of latest Iphone and Samsung in australia",
         llm=ChatOpenAI(model="gpt-4.1-nano-2025-04-14"),
         enable_memory=False,
-        proxy=utils.PROXY(country_code='au', city='brisbane', session_time=10),
+        proxy=stealth_utils.PROXY(country_code='au', city='brisbane', session_time=10),
     )
     await agent.run()
